Add support for RTC power off control
The RTC power off control is configured through the device tree.
To use the RTC power off feature, the "rtc-pwrctl-enable" property
must be set to <1>.

If the "rtc-power-off-time" is masked, the RTC power off control
feature will not be enabled. By default, BSP does not enable this
feature. If "rtc-power-off-time" exists and its value is between 4 and
7 inclusive, the RTC power off feature will be activated.

The "rtc-power-off-time" determines the duration in seconds that the
RTC_nWake key must be held to trigger RTC power off.

Change-Id: Ibebe071726946facdfc2d694eb64d763eab61104